make the ej header the default theme
