+++ themes/brasil-que-o-povo-quer/header/header.component.scss
@@ -1,0 +1,324 @@
+// Site header styles
+// --------------------------------------------------
+
+@import '../../../node_modules/bootstrap/scss/bootstrap';
+@import '../../../src/variables';
+@import '../../../src/mixins';
+
+.menu-icon{ display: none; }
+
+.fix-height{
+    width: 100%;
+    height: 40px;
+    display: none;
+}
+
+.site-header.unlogged{
+    @extend .clearfix;
+    
+    max-width: 1140px;
+    margin: 0 auto;
+    margin-top: 20px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+
+
+    @media screen and (min-width: 640px) and (max-width: 1200px){  
+        .site-links, .access{
+            margin-right: 20px !important;
+            margin-left: 0px !important;
+        }
+    }
+
+    @include media-breakpoint-down(sm) {  
+        flex-direction: column;
+        justify-items: center;
+
+        .fix-height{ display: block; }    
+
+        .menu-icon{
+            display: block;
+            position: absolute;
+            left: 17px;
+            top: 17px;
+            img { width: 28px !important; height: auto !important; } 
+        }
+
+        .site-logo, .access, .site-links{
+            margin: 0px !important;
+            padding-left: 0px !important;
+        }
+
+        .site-logo{
+            position: fixed;
+            background: white;
+            z-index: 100;
+            width: 100%;
+            text-align: center;
+            top: 0;
+            box-shadow: 0 0 5px #797979;
+            height: 60px;
+            padding-top: 5px;
+        }
+
+        .without-profile{
+            position: fixed;
+            top: 60px;
+            left: -100%;
+            width: 50vw;
+            background-color: #eae6e5;
+            height: 100%;
+            z-index: 10;
+            box-shadow: 0 0 5px #6f6f6f;
+            display: flex;
+            flex-direction: column-reverse;
+            justify-content: flex-end;
+            padding: 20px 10px;
+            transition: left .6s;
+
+            .access, .site-links{
+                display: flex;
+                flex-direction: column;
+    
+                li{ 
+                    margin-left: 0 !important; 
+                    text-align: center;
+                    padding: 10px 0px;
+                }
+            }
+
+            .site-links li { padding: 20px 0px; }
+        }
+
+    }        
+
+    &:after{
+        display: none !important;
+    }
+
+    .site-logo{
+        float: left;
+        @extend .ml-5;
+    }
+
+    .without-profile{
+        
+        .access{
+            @extend .row;
+            font-size: small;
+            color: #F10000;
+            
+            @include media-breakpoint-down(md) {  
+                a { line-height: 2.3; }
+            }
+        }
+        .site-links{
+            @extend .row;
+            font-size: small;
+            a{
+                color: black;
+            }
+        }
+        
+        @extend .text-uppercase;
+        ul{
+            list-style: none;
+            @extend .justify-content-end;
+            li{
+                @extend .ml-4;
+                display: inline;
+                @extend .align-middle;
+                a{
+                    cursor: pointer;
+                }
+            }
+        }
+    }  
+}
+
+.site-logo  img{
+    @include media-breakpoint-down(sm) {  
+        height: 40px;
+    }
+}      
+
+.logged.site-header {
+    background-color: $site-header-bg-color;
+    box-shadow: 2px 2px 4px 0 rgba(0, 0, 0, 0.5);
+    height: 60px;
+    @include transition(width 0.5s ease);
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    z-index: 1010;
+    @include media-breakpoint-up(sm) {
+        height: 80px;
+    }
+
+    .site-logo {
+        height: 60px;
+        width: 100%;
+        border-left: 1px solid black;
+        background-color: rgba(255, 255, 255, 0.1);
+        
+        @include media-breakpoint-up(sm) {
+            height: 80px;
+            left: 100px;
+        }
+        h1 {
+            margin: 0;
+            height: 100%;
+            width: 100%;
+        }
+        a {
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            height: 100%;
+        }
+    }
+
+    .user-info {
+        box-sizing: border-box;
+        color: #000;
+        cursor: pointer;
+        height: 60px;
+        padding: 6px;
+        box-shadow: 0 1px 2px 0 rgba(0,0,0,0.5);
+        float: left;
+        width: 54px;
+        z-index: 1;
+        background-color: white;
+        width: 100px;
+        @include media-breakpoint-up(xs) {
+            width: 85px;
+        }
+
+        @include media-breakpoint-up(sm) {
+            height: 80px;
+            padding: 10px;
+            width: 100px;
+        }
+
+        .user-icon {
+            border-radius: 100%;
+            // box-shadow: 2px 2px 4px 0 rgba(0, 0, 0, 0.5);
+            float: none;
+            height: 34px;
+            margin: 0 auto;
+            text-align: center;
+            width: 34px;
+            @include media-breakpoint-up(sm) {
+                height: 40px;
+                width: 40px
+            }
+            img {
+                height: 34px;
+                width: 34px;
+                @include media-breakpoint-up(sm) {
+                    height: 40px;
+                    width: 40px
+                }
+            }
+        }
+        .user-name {
+            display: none;
+            float: left;
+            font-size: $font-size-base;
+            font-weight: $semi-bold;
+            margin: 15px 0 0 10px;
+            text-align: left;
+            color: #D93E25;
+            max-width: 160px;
+            word-wrap: break-word;
+            @include media-breakpoint-up(xs) {
+                margin: 0px 0 0 10px;
+            }
+
+        }
+        .menu-label, .hide-menu {
+            display: block;
+            font-size: 1.6vmax;
+            font-weight: $bold;
+            margin-top: 4px;
+            text-align: center;
+            color: #D93E25;
+            @include media-breakpoint-up(xs) {
+                font-size: 10px;
+            }
+            @include media-breakpoint-up(sm) {
+                font-size: $font-size-sm;
+                margin-top: 5px;
+            }
+        }
+        .hide-menu {
+            display: none;
+        }
+    }
+}
+
+.logged.site-header .category-page.site-logo a,
+.unlogged.site-header .category-page.site-logo a {
+  justify-content: left;
+
+  img {
+    width: 120px;
+    margin: 10px;
+  }
+}
+
+.logged.site-header .category-page.site-logo,
+.unlogged.site-header .category-page.site-logo,
+.logged.site-header .category-page.site-logo h1,
+.unlogged.site-header .category-page.site-logo h1 {
+  width: 120px;
+}
+
+.header-category {
+  position: absolute;
+  top: 50%;
+  margin-top: -24px;
+  left: 0;
+  width: 100%; 
+  line-height: 48px;
+  text-align: center;
+  cursor: pointer;
+  outline: none;
+  z-index: -1;
+
+  h2 {
+    color: #4a4a4a;
+    font-size: 20px;
+    text-transform: uppercase;
+    font-weight: 500;
+    line-height: 48px;
+    margin: 0;
+    margin-left: 16px;
+    padding: 0;
+    display: inline;
+  }
+
+  img {
+    height: 48px;
+    box-shadow: 0 0 4px 0 rgba(0, 0, 0, 0.5);
+  }
+}
+            
+@include media-breakpoint-down(sm) {
+  .site-header.category-page {
+    height: 120px !important;
+  }
+
+  .header-category {
+    top: 63px;
+    margin-top: 0;
+    padding-left: 95px;
+    white-space: nowrap;
+  }
+
+  .category-page .user-info {
+    height: 120px !important;
+  }
+}
