--- themes/brasil-que-o-povo-quer/header/header.html
@@ -1,1 +1,0 @@
-<p>Theme BRASIL QUE O POVO QUER </p>
