--- themes/brasil-que-o-povo-quer/header/header.scss
@@ -1,3 +1,0 @@
-p {
-  color: blue;
-}
