+++ themes/theme-default/header/header.component.html
@@ -1,43 +1,1 @@
-<header class="site-header" [ngClass]="{ 'logged': isLogged(), 'unlogged': !isLogged(), 'category-page': !!getStyles() }">
-    <div *ngIf="profile" class='with-profile'>
-        <div class="user-info" (click)="toggleMenu()"  >
-            <div class="user-icon">
-                <img src="/assets/images/icons/menu_icon.svg" alt="Some user" class="img-circle">
-                <div class="menu-label">{{ 'header.menu' | translate }}</div>
-                <div class="hide-menu">{{ 'header.hide' | translate }}</div>
-            </div>
-            <div class="user-name">{{profile.name | shorten: 18: '...'}}</div>
-        </div>
-        <app-notifications [profile]="profile" ></app-notifications>
-        
-    </div>
-    <div class="site-logo" [ngClass]="{ 'category-page': !!getStyles() }">
-        <h1><a routerLink="/"><img src="/assets/images/logo.svg" alt="{{ 'header.site' | translate }}"></a></h1>
-
-        <span *ngIf="category" class="header-category" routerLink="/{{category.slug}}">
-          <img [src]="category.image" />
-          <h2>{{category.name}}</h2>
-        </span>
-
-        <div class="menu-icon" (click)="toggleMenu()">
-            <img src="/assets/images/icons/menu-icon.png" alt="">
-        </div>
-    </div>
-    <div *ngIf="!profile" class="without-profile">
-        <ul class='access'>
-            <li><a (click)="openRegistration()">{{ 'header.registration' | translate }}</a></li>
-            <li>
-                <a (click)="openLogin()">{{ 'header.login' | translate }}</a>
-                <i  class="fa fa-user"></i>
-            </li>
-        </ul>
-        <ul class='site-links'>
-            <li><a (click)="loginWithFacebook()">Logar com Facebook</a></li>
-            <li><a (click)="loginWithTwitter()">Logar com Twitter</a></li>
-            <li><a routerLink="/sobre-nos">Sobre Nós</a></li>
-            <li><a routerLink="/conversations">Conversas</a></li>
-            <li><a routerLink="/perguntas-frequentes">Perguntas Frequentes</a></li>
-        </ul>
-    </div>
-    <div *ngIf="!isLogged()" class="fix-height"></div>
-</header>
+<p>Theme EJ HEADER </p>
